Renamed "tree" to "navigationManager
