+++ library/src/scripts/components/icons/navigationManager.tsx
